Separate Path structure
